Delete Barcodes Changes

+++ admin/applications/products/actions/deleteBarcode.php
@@ -4,16 +4,30 @@
 		if ($Barcode->status == 'O'){
 			$response = array(
 				'success' => true,
-				'errorMsg' => 'That barcode is currently rented out, please return it before deleting.'
+				'errorMsg' => sysLanguage::get('TEXT_BARCODE_OUT')
 			);
 		}elseif ($Barcode->status == 'P'){
 			$response = array(
 				'success' => true,
-				'errorMsg' => 'That barcode has been purchased, it cannot be deleted.'
+				'errorMsg' => sysLanguage::get('TEXT_BARCODE_PURCHASED')
 			);
 		}else{
-			$Barcode->delete();
-			$response = array('success' => true);
+			$Qproducts = Doctrine_Query::create()
+			->from('ProductsInventoryBarcodes pib')
+			->leftJoin('pib.OrdersProductsReservation opr')
+			->where('pib.barcode_id=?', $_GET['bID'])
+			->andWhere('opr.start_date >= ?', date('Y-m-d'))
+			->execute(array(), Doctrine_Core::HYDRATE_ARRAY);
+
+			if(count($Qproducts) > 0){
+				$response = array(
+					'success' => true,
+					'errorMsg' => sysLanguage::get('TEXT_FUTURE_RESERVATION')
+				);
+			}else{
+				$Barcode->delete();
+				$response = array('success' => true);
+			}
 		}
 	}else{
 		$response = array('success' => false);
